Added math utility class. Added RNG class. Refactored Need/Job. Turned most logic-related game object properties into attributes and made a generic base class for attribute storage and retrieval. Added Move job and Wander job. Fixed Box bug.

+++ src/logic/WorldUpdater.hpp
@@ -12,6 +12,6 @@ class WorldUpdater
 public:
     WorldUpdater(World &world);
 
-    void update(double dt);
+    void update(unsigned int dt);
 };
 // LCOV_EXCL_STOP
